fixed validation bug
