+++ src/components/EndpointTestComponent/JestMatchers.js
@@ -1,0 +1,37 @@
+const jestMatchers = [
+  '',
+  'to Be',
+  'to Equal (object)',
+  'to Have Been Called',
+  'to Have Been Called Times (number)',
+  'to Have Been Called With (arg1,...)',
+  'to Have Been Last Called With (arg1,...)',
+  'to Have Been Nth Called With (nth call, arg1,...)',
+  'to Have Length (number)',
+  'to Have Property (keyPath, value[optional])',
+  'to Be Close To (number, number of digits[optional])',
+  'to Be Defined',
+  'to Be Undefined',
+  'to Be Falsy',
+  'to Be Truthy',
+  'to Be NaN',
+  'to Be Greater Than (number)',
+  'to Be Greater Than Or Equal (number)',
+  'to Be Less Than (number)',
+  'to Be Less Than Or Equal (number)',
+  'to Be Instance Of (Class)',
+  'to Contain (item in an array)',
+  'to Contain Equal (an object in an array)',
+  'to Match (regexp or string)',
+  'to Match Object (object)',
+  'to Srict Equal (object)',
+  'to Throw (error[optional])',
+];
+
+//for mock fuctions only:
+//   'to Have Returned',
+//   'to Have Returned __ Times (number)',
+//   'to Have Last Returned With',
+// ];
+
+export default jestMatchers;
